AJTEST-492 Fix windows build of advtdiscov
Core codebase moved to /W4 warning level for Windows, from /W2.
This is a more stricter warning setting and the test programs need
to be updated accordingly.

Q: But, why do the test programs even need an update when such a
   thing changes in the Core codebase?
A: Currently, the test programs use the same SConscript files as Core
   and this tight coupling is what forces the test programs to be
   updated as well.
   The right thing is to uncouple these two. But for now, the quicker
   approach is to follow suit and update the test programs as well.

Change-Id: Ib783733633a8749726c5529bc8c2d28282e8ae0a
Signed-off-by: Praveen Bhamidipati <praveenb@qce.qualcomm.com>