ASACORE-2800 Fix ajtcsctest.exe failure
Fix SecurityClaimApplicationTest.no_state_signal_after_update_identity failure.

A recent change to Keystore caused this test to fail because key expiration in
the previous version of keystore was not reliable. This particular test had a
security manager BA along with an "admin" BA sharing the same Keystore; when
they are communicated over ECDHE_NULL, the peer secret would get expired right
away causing the second BA to no longer be able to access the GUID.

This change gets rid of the "admin" BA and just use the security manager BA to
generate the identity cert. The test to ensure that there's no signal after
UpdateIdentity() is called now passes.

I also tested other tests in ajtcsctest to ensure that they pass.

Change-Id: Ic887f74dfd7a2253c43ac8ed029c7fcc9ae60f69
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>