ASACORE-2775 ASACORE-2912 Fix broken test and test manifest signing
Broken test was setting a simple policy, and because trust anchors
are computed based on policy set, received manifests would no longer
validate and test would break. Fix is to extract the entries from
the default policy that set trust anchors and add them to the
permissive policy set by the test.

Change test methods that sign methods to use new core methods in
PermissionConfigurator, instead of duplicating that functionality.

Change-Id: I41e7c91d85c594c303515d6ce8edfd8aa182a41b
Signed-off-by: Kevin Kane <kkane@microsoft.com>