ASACORE-3048 Make ajtcsctest (interaction test) build correctly in debug mode
Crtdbg.h relies on _DEBUG macro being defined, which is the foundation of some
of the other macros such as _ASSERTE, AJ_VERIFY, AJ_ASSERT, etc. This works
properly in AllJoyn Core repo because windows\mapping.h manually defines
_ASSERTE, which is not needed if the code is compiled in Visual Studio.

This is making the equivalent change for ajtcsctest, so that AJ_VERIFY and
AJ_ASSERT don't get compiled out in debug mode (Windows only).

Change-Id: I481e7c6d855a3d0add84e9d665fd74aa2f6e555d
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>