ASACORE-3319 Fix a bug in bbtest in stdin handling
The issue is bbtest is using Socket API select() for stdin (to test
user's inputs), which is not supported on Windows. This fix is to
implement the equivalent functionality ("read from stdin with
timeout") as well as fix a number of the existing buffer overruns
that could be encountered if user's input is too long.

Change-Id: I8d567234a9f0a1672747b565596569c97254d51c
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>